+++ changelog.md
@@ -275,7 +275,9 @@
 - Added `genasts.genAst` that avoids the problems inherent with `quote do` and can
   be used as a replacement.
 
-- Added `std/syntaxes` to hint syntax highlighters about string literals.
+- Added `stdx/` for standard library extensions, see `stdx/readme.md`
+
+- Added `stdx/lang_prefixes` to help tooling (e.g. syntax highlighters) how to interpret string literals.
 
 ## Language changes